Move common APIs out from basic_tests.ts
